Add files via upload
Added a WIP section (commented out) that takes the value of $manager, fetches an array of all users matching the input string, then generates a numbered list of users to be targeted and chosen to be the official value of $manager. 
Began a section that will generate a "note" to be copy/pasted into the ConnectWise Manage onboarding ticket detailing the work done.