Add files via upload
+ Added small summary at the top to TLDR what the script does.
+ Corrected login sections to pull password from OneDrive file.
o Moved WIP manager verification code to end of script.
+ Reformatted the script to make it easier to read.
o This script might be moved into script 'LandryLabs.ps1' inside of a function to be called there.

12/19
+ Added an 'onboarding summary note' text file output (in LandryLabs\Tasks folder) to copy/paste to the CWM ticket.
o Starting to work on the feature that will send emails. This will start as direct email sending with no access to mailbox folders.

12/20
+ Now uses $department and $practice data to detect&confirm the OU the user object will be moved to.
+ Now moves the user automatically to their intended OU to be in AADSync scope.
+ Adjusted the mirror manager line of the ConnectwiseManage note to conditionally show 'user mirrored to X" only if ($mirroruser -ne "N") and otherwise states that no mirror user was assigned.
+ Adjusted the CWMNote's generated text file name to be generically "Onboarding_Logs" and to append new logs into that one file and added a timestamp at the beginning.
+ Corrected email showing blank in the CWMNotes.

12/22
o Located the timestamp for AAD's last directory sync and added commented-out code for fetching it to consider implementing that information in some way with the 'waiting for AADSync' output statements for better monitoring.
o Will look into ways to copy all visible output of the script to the onboarding log text file.
+ Confirmed Contract Labor security group is now been added successfully.

12/22
+ Contract Labor security group is now been added successfully.
+ An email alert of task completion is now successfully sent to daniel.landry@sparkhound.com
+ Created account/mailbox for 'Landrylabs.bot' assistant.
+ Adjusted the 'task complete' email alert to push from landrylabs.bot@sparkhound.com to daniel.landry@sparkhound.com instead of directly to daniel.landry@sparkhound.com.
+ Created a second alert containing the email address and temp pass of the user and pushes it to hr@sparkhound.com.
+ Adjusted AzureAD connection to successfully login automatically.

12/23
+ Added small summary at the top to TLDR what the script does.
+ Corrected login sections to pull password from OneDrive file.
o Moved WIP manager verification code to end of script.
+ Reformatted the script to make it easier to read.
o This script might be moved into script 'LandryLabs.ps1' inside of a function to be called there.